proxy: move http related thing to httpproxy
